Add files via upload
In this project, we are creating a web application through which users can summarise their files. The input can be a link containing text, audio, and video files, or it can be normal text, audio, and video files. Automatic text summarization is basically a summary of the given paragraph using natural language processing and machine learning. We review the different processes for summarization and describe the effectiveness and shortcomings of the different methods.